Fixed data collation bug with float types

+++ src/main/java/com/bt/betalab/dasboardservice/api/SimulationData.java
@@ -18,26 +18,26 @@ public class SimulationData {
     private int bouncedCalls;
     private int resolvedCalls;
     private int unresolvedCalls;
-    private long callDelay;
+    private float callDelay;
 
-    private long overallWaitTime;
-    private long averageWaitTime;
-    private long longestWaitTime;
-    private long shortestWaitTime;
+    private float overallWaitTime;
+    private float averageWaitTime;
+    private float longestWaitTime;
+    private float shortestWaitTime;
 
-    private long overallServiceTime;
-    private long averageServiceTime;
-    private long longestServiceTime;
-    private long shortestServiceTime;
+    private float overallServiceTime;
+    private float averageServiceTime;
+    private float longestServiceTime;
+    private float shortestServiceTime;
 
-    private long predictedHappinessSum;
-    private long averagePredictedHappiness;
+    private float predictedHappinessSum;
+    private float averagePredictedHappiness;
 
-    private long actualHappinessSum;
-    private long averageActualHappiness;
+    private float actualHappinessSum;
+    private float averageActualHappiness;
 
-    private long easySum;
-    private long easyFraction = 0;
+    private float easySum;
+    private float easyFraction = 0;
     private List<CallData> calls;
 
     public String getSimulationId() {
@@ -88,59 +88,59 @@ public class SimulationData {
         this.resolvedCalls = resolvedCalls;
     }
 
-    public long getCallDelay() {
+    public float getCallDelay() {
         return callDelay;
     }
 
-    public void setCallDelay(long callDelay) {
+    public void setCallDelay(float callDelay) {
         this.callDelay = callDelay;
     }
 
-    public long getAverageWaitTime() {
+    public float getAverageWaitTime() {
         return averageWaitTime;
     }
 
-    public void setAverageWaitTime(long averageWaitTime) {
+    public void setAverageWaitTime(float averageWaitTime) {
         this.averageWaitTime = averageWaitTime;
     }
 
-    public long getLongestWaitTime() {
+    public float getLongestWaitTime() {
         return longestWaitTime;
     }
 
-    public void setLongestWaitTime(long longestWaitTime) {
+    public void setLongestWaitTime(float longestWaitTime) {
         this.longestWaitTime = longestWaitTime;
     }
 
-    public long getShortestWaitTime() {
+    public float getShortestWaitTime() {
         return shortestWaitTime;
     }
 
-    public void setShortestWaitTime(long shortestWaitTime) {
+    public void setShortestWaitTime(float shortestWaitTime) {
         this.shortestWaitTime = shortestWaitTime;
     }
 
-    public long getAveragePredictedHappiness() {
+    public float getAveragePredictedHappiness() {
         return averagePredictedHappiness;
     }
 
-    public void setAveragePredictedHappiness(long averagePredictedHappiness) {
+    public void setAveragePredictedHappiness(float averagePredictedHappiness) {
         this.averagePredictedHappiness = averagePredictedHappiness;
     }
 
-    public long getAverageActualHappiness() {
+    public float getAverageActualHappiness() {
         return averageActualHappiness;
     }
 
-    public void setAverageActualHappiness(long averageActualHappiness) {
+    public void setAverageActualHappiness(float averageActualHappiness) {
         this.averageActualHappiness = averageActualHappiness;
     }
 
-    public long getEasyFraction() {
+    public float getEasyFraction() {
         return easyFraction;
     }
 
-    public void setEasyFraction(long easyFraction) {
+    public void setEasyFraction(float easyFraction) {
         this.easyFraction = easyFraction;
     }
 
@@ -160,67 +160,67 @@ public class SimulationData {
         this.unresolvedCalls = unresolvedCalls;
     }
 
-    public long getOverallWaitTime() {
+    public float getOverallWaitTime() {
         return overallWaitTime;
     }
 
-    public void setOverallWaitTime(long overallWaitTime) {
+    public void setOverallWaitTime(float overallWaitTime) {
         this.overallWaitTime = overallWaitTime;
     }
 
-    public long getOverallServiceTime() {
+    public float getOverallServiceTime() {
         return overallServiceTime;
     }
 
-    public void setOverallServiceTime(long overallServiceTime) {
+    public void setOverallServiceTime(float overallServiceTime) {
         this.overallServiceTime = overallServiceTime;
     }
 
-    public long getAverageServiceTime() {
+    public float getAverageServiceTime() {
         return averageServiceTime;
     }
 
-    public void setAverageServiceTime(long averageServiceTime) {
+    public void setAverageServiceTime(float averageServiceTime) {
         this.averageServiceTime = averageServiceTime;
     }
 
-    public long getLongestServiceTime() {
+    public float getLongestServiceTime() {
         return longestServiceTime;
     }
 
-    public void setLongestServiceTime(long longestServiceTime) {
+    public void setLongestServiceTime(float longestServiceTime) {
         this.longestServiceTime = longestServiceTime;
     }
 
-    public long getShortestServiceTime() {
+    public float getShortestServiceTime() {
         return shortestServiceTime;
     }
 
-    public void setShortestServiceTime(long shortestServiceTime) {
+    public void setShortestServiceTime(float shortestServiceTime) {
         this.shortestServiceTime = shortestServiceTime;
     }
 
-    public long getPredictedHappinessSum() {
+    public float getPredictedHappinessSum() {
         return predictedHappinessSum;
     }
 
-    public void setPredictedHappinessSum(long predictedHappinessSum) {
+    public void setPredictedHappinessSum(float predictedHappinessSum) {
         this.predictedHappinessSum = predictedHappinessSum;
     }
 
-    public long getActualHappinessSum() {
+    public float getActualHappinessSum() {
         return actualHappinessSum;
     }
 
-    public void setActualHappinessSum(long actualHappinessSum) {
+    public void setActualHappinessSum(float actualHappinessSum) {
         this.actualHappinessSum = actualHappinessSum;
     }
 
-    public long getEasySum() {
+    public float getEasySum() {
         return easySum;
     }
 
-    public void setEasySum(long easySum) {
+    public void setEasySum(float easySum) {
         this.easySum = easySum;
     }
 }
